Fix folds all opening on first za
When folding methods with za, the first time the command ran, all the
folds would collapse. I'd then have to zR to open everything back up and
it would work great from then on.

Turns out it was a quirk with the default foldlevelstart being -1. By
setting it to 99, folding works as expected.